requirements.txt
ultralytics>=8.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.9
gradio>=4.0.0
opencv-python>=4.8.0
pdf2image>=1.17.0
scikit-image>=0.22.0
